Update DBUtil.java
Define java class DBUtil

This method fetches a database connection. It ensures the connection is open by checking if it's null or closed. Then, it fetches connection details from a properties file, loads the database driver, and establishes a connection using the obtained details. Finally, it returns the connection object.