Update CartServiceImpl.java
Creating a method called removeAProduct().
This method contains remove a product from the cart.
If the user remove a product from the cart. It will check the userid and productid. Then it will execute SQL statement to remove the product. 
If the executeupdate() is greater than 0. It means the products was removed.