Update CartServiceImpl.java
Creating a method for updating product to cart.

In here, It updates the quantity of a product in the user's cart based on the provided user ID, product ID, and new quantity.
It first checks if the product is already in the user's cart by executing a SQL query.
If the product is found in the cart, the new quantity is greater than 0. It updates the quantity of the product in the cart using an SQL UPDATE statement.
If the new quantity is 0, it deletes the product from the cart using an SQL DELETE statement.
If the product is not found in the cart. It inserts a new entry into the cart with the provided quantity using an SQL INSERT statement.