Update OrderServiceImpl.java
This OrderServiceImpl class implements the OrderService interface.

This method takes the username and the paid amount as input and attempts to place an order based on the user's cart items. It returns a success message ("Order Placed Successfully!") if successful, or a failure message ("Order Placement Failed!")

It retrieves all the items in the user's cart using the “getAllCartItems” method from “CartServiceImpl”. If the cart is empty, the method returns with the status indicating failure.

A “TransactionBean” object is created with the user's username and the paid amount.

For each item in the cart it calculates the total amount and creates an “OrderBean” object and attempts to add the order. If any step fails (adding order, removing product from the cart, or updating product quantity) the loop breaks and the method returns with the status indicating failure.

If all orders are successfully processed, it attempts to add the transaction. If successful it sends a transaction success message to the user and updates the status to "Order Placed Successfully!".