Update MailMessage.java
This code sends an email to a user who has successfully registered on the "Green Life" platform, welcoming them and providing information about the platform's services and contact details.
  This method takes two parameters as emailId (the email address of the recipient) and name (the name of the user).
It constructs an HTML formate message to be sent as an email. The message includes a greeting, congratulations message for successful registration, information about the platform's services, contact information, and a closing note.
The HTML message is stored in the htmlTextMessage variable.
It then attempts to send the email using a method called sendMail from a class named JavaMailUtil. This implies that there is another class called JavaMailUtil containing the necessary functionality to send emails.