Update MailMessage.java
This method is used to notify customers about when the product get available in the store. 

This method take parameters as recipientEmail, name, prodName, and prodId.

It constructs an HTML email message to notify the recipient that a product (prodName) with the given ID (prodId) is now available at the store.

The subject of the email is constructed dynamically based on the product name.

The body of the email contains a personalized message addressing the recipient by name and informing them that the product they were previously interested in is now available.

If any error occur the exception will executed.