Update UserService.java
This package has some set of methods.

registerUser(String userName, Long mobileNo, String emailId, String address, int pinCode, String password)- This method to register a user with various parameters like username, mobile number, email, address, PIN code, and password. It likely returns a message or status indicating the result of the registration process.

registerUser(UserBean user)- This method seems be an overloaded version of the previous registerUser method, where it accepts a UserBean object instead of individual parameters.

isRegistered(String emailId)- This method checks if a user with the specified email ID is already registered. It returns a boolean value indicating whether the user is registered or not.

isValidCredential(String emailId, String password)- This method checks if the provided email ID and password combination is valid. It likely returns a message or status indicating the validity of the credentials.

getUserDetails(String emailId, String password)- This method retrieves details of a user based on the provided email ID and password. It probably returns a UserBean object containing user details.

getFName(String emailId)- This method retrieves the first name of the user based on the provided email ID.

getUserAddr(String userId)-  This method retrieves the address of the user based on the provided user ID. However, there seems to be a discrepancy here; it's named getUserAddr but takes a userId parameter, which suggests it should actually get the address based on user ID, not email ID as the name implies.